pyproject.toml
# Ultralytics MkDocs plugin 🚀, AGPL-3.0 license

# Overview:
# This pyproject.toml file manages the build, packaging, and distribution of the MkDocs Ultralytics Plugin.
# It defines essential project metadata, dependencies, and settings used to develop and deploy the plugin.

# Key Sections:
# - [build-system]: Specifies the build requirements and backend (e.g., setuptools, wheel).
# - [project]: Includes details like name, version, description, authors, dependencies, and more.
# - [project.optional-dependencies]: Provides additional, optional packages for extended features.
# - [tool.*]: Configures settings for various tools (e.g., pytest) used in the project.

# Installation:
# The MkDocs Ultralytics Plugin can be installed using the command: 'pip install mkdocs-ultralytics-plugin'
# For development purposes, you can install the package in editable mode with: 'pip install -e .'
# This approach allows for real-time code modifications without the need for re-installation.

# Documentation:
# For comprehensive documentation and usage instructions, visit: https://github.com/ultralytics/mkdocs

[build-system]
requires = ["setuptools>=57.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mkdocs-ultralytics-plugin"
version = "0.1.6"
description = "An MkDocs plugin that provides Ultralytics Docs customizations at https://docs.ultralytics.com."
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
keywords = ["Ultralytics", "Docs", "MkDocs", "Plugin"]  # Optional
authors = [
    { name = "Glenn Jocher", email = "glenn.jocher@ultralytics.com" }  # Optional
]
maintainers = [
    { name = "Glenn Jocher", email = "glenn.jocher@ultralytics.com" }  # Optional
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
]  # Optional, for a list of valid classifiers, see https://pypi.org/classifiers/
dependencies = [
    "mkdocs>=1.0",
    "beautifulsoup4>=4.9.3",
    "pyyaml",
    "requests>=2.31.0",
]

[project.urls]  # Optional
"Bug Reports" = "https://github.com/ultralytics/mkdocs/issues"
"Funding" = "https://ultralytics.com"
"Source" = "https://github.com/ultralytics/mkdocs/"

[project.entry-points."mkdocs.plugins"]
ultralytics = "plugin:MetaPlugin"  # executes the function `MetaPlugin` from this package when "plugin" is called.

[project.optional-dependencies]  # Optional
dev = ["pytest"]

[tool.setuptools]  # This is configuration specific to the `setuptools` build backend.
package-data = { "sample" = ["*.yaml"] }
